move message extraction logic
