+++ src/app/views/app-sections/StatusMessages.tsx
@@ -1,48 +1,17 @@
-import { Link, MessageBar } from '@fluentui/react';
-import { Fragment } from 'react';
+import { MessageBar } from '@fluentui/react';
 import { useDispatch } from 'react-redux';
 
 import { AppDispatch, useAppSelector } from '../../../store';
 import { IQuery } from '../../../types/query-runner';
 import { setSampleQuery } from '../../services/actions/query-input-action-creators';
 import { clearQueryStatus } from '../../services/actions/query-status-action-creator';
-import { GRAPH_URL } from '../../services/graph-constants';
-import {
-  convertArrayToObject, extractUrl, getMatchesAndParts,
-  matchIncludesLink, replaceLinks
-} from '../../utils/status-message';
 import { translateMessage } from '../../utils/translate-messages';
+import MessageDisplay from '../common/MessageDisplay';
 
 const StatusMessages = () => {
   const dispatch: AppDispatch = useDispatch();
   const { queryRunnerStatus, sampleQuery } =
     useAppSelector((state) => state);
-
-  function displayStatusMessage(message: string, urls: any) {
-    const { matches, parts } = getMatchesAndParts(message);
-
-    if (!parts || !matches || !urls || Object.keys(urls).length === 0) {
-      return message;
-    }
-
-    return parts.map((part: string, index: number) => {
-      const includesLink = matchIncludesLink(matches, part);
-      const displayLink = (): React.ReactNode => {
-        const link = urls[part];
-        if (link) {
-          if (link.includes(GRAPH_URL)) {
-            return <Link onClick={() => setQuery(link)} underline>{link}</Link>;
-          }
-          return <Link target="_blank" href={link} underline>{link}</Link>;
-        }
-      };
-      return (
-        <Fragment key={part + index}>{includesLink ?
-          displayLink() : part}
-        </Fragment>
-      );
-    })
-  }
 
   function setQuery(link: string) {
     const query: IQuery = { ...sampleQuery };
@@ -54,20 +23,14 @@ const StatusMessages = () => {
 
   if (queryRunnerStatus) {
     const { messageType, statusText, status, duration, hint } = queryRunnerStatus;
-    let urls: any = {};
-    let message = status.toString();
-    const extractedUrls = extractUrl(status.toString());
-    if (extractedUrls) {
-      message = replaceLinks(status.toString());
-      urls = convertArrayToObject(extractedUrls);
-    }
+
 
     return <MessageBar messageBarType={messageType}
       isMultiline={true}
       onDismiss={() => dispatch(clearQueryStatus())}
       dismissButtonAriaLabel='Close'
       aria-live={'assertive'}>
-      {`${statusText} - `}{displayStatusMessage(message, urls)}
+      {`${statusText} - `}{<MessageDisplay message={status.toString()} onSetQuery={setQuery} />}
 
       {duration && <>
         {` - ${duration} ${translateMessage('milliseconds')}`}
